# tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"Pick_UserOutput.Exclude_keyofUserOutput.password__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string",
						"nullable": true
					},
					"avatar_url": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"is_owner": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"email",
					"firstname",
					"lastname",
					"avatar_url",
					"phone",
					"created_at",
					"is_active",
					"is_owner"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UserOutput.password_": {
				"$ref": "#/components/schemas/Pick_UserOutput.Exclude_keyofUserOutput.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UserRecord": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_UserOutput.password_"
					},
					{
						"properties": {
							"avatar_url": {
								"type": "string"
							}
						},
						"required": [
							"avatar_url"
						],
						"type": "object"
					}
				]
			},
			"UserCredentials": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"remember": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"%24Result.DefaultSelection_Prisma.%24UserPayload_": {
				"properties": {
					"is_owner": {
						"type": "boolean"
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"avatar_url": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"is_owner",
					"is_active",
					"created_at",
					"password",
					"phone",
					"avatar_url",
					"lastname",
					"firstname",
					"email",
					"id"
				],
				"type": "object"
			},
			"User": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24UserPayload_",
				"description": "Model User"
			},
			"UserOutput": {
				"$ref": "#/components/schemas/User"
			},
			"Pick_User.email-or-firstname-or-lastname-or-password-or-is_owner_": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string",
						"nullable": true
					},
					"is_owner": {
						"type": "boolean"
					}
				},
				"required": [
					"password",
					"email",
					"firstname",
					"lastname",
					"is_owner"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreateInput": {
				"$ref": "#/components/schemas/Pick_User.email-or-firstname-or-lastname-or-password-or-is_owner_"
			},
			"Partial_Pick_UserUpdateInputWithOldPassword.old_password-or-email-or-firstname-or-lastname-or-password-or-phone-or-avatar_url__": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string",
						"nullable": true
					},
					"avatar_url": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"old_password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserUpdateInput": {
				"$ref": "#/components/schemas/Partial_Pick_UserUpdateInputWithOldPassword.old_password-or-email-or-firstname-or-lastname-or-password-or-phone-or-avatar_url__"
			},
			"Partial_CategoryOutput_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"restaurant_id": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CategoryProps": {
				"$ref": "#/components/schemas/Partial_CategoryOutput_"
			},
			"WithIsUsed_CategoryProps_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/CategoryProps"
					},
					{
						"properties": {
							"is_used": {
								"type": "boolean"
							}
						},
						"required": [
							"is_used"
						],
						"type": "object"
					}
				]
			},
...
			"Category": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24CategoryPayload_",
				"description": "Model Category"
			},
			"CategoryOutput": {
				"$ref": "#/components/schemas/Category"
			},
			"Pick_CategoryOutput.label_": {
				"properties": {
					"label": {
						"type": "string"
					}
				},
				"required": [
					"label"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"CategoryCreateInput": {
				"$ref": "#/components/schemas/Pick_CategoryOutput.label_"
			},
...
			"WithHide_Partial_Pick_CategoryProps.label___": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_Pick_CategoryProps.label__"
					},
					{
						"properties": {
							"hide": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
			"CategoryUpdateInput": {
				"$ref": "#/components/schemas/WithHide_Partial_Pick_CategoryProps.label___"
			},
...
			"Dishes": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24DishesPayload_",
				"description": "Model Dishes"
			},
			"DishOutput": {
				"$ref": "#/components/schemas/Dishes"
			},
			"DishesByCategoryOutput": {
				"properties": {
					"dishes": {
						"items": {
							"$ref": "#/components/schemas/DishOutput"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					}
				},
				"required": [
					"dishes",
					"category"
				],
				"type": "object"
			},
			"Prisma.Decimal": {
				"type": "string"
			},
			"Pick_DishOutput.name-or-photo_url-or-allergen-or-price-or-description_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"photo_url": {
						"type": "string",
						"nullable": true
					},
					"allergen": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"$ref": "#/components/schemas/Prisma.Decimal"
					},
					"description": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name",
					"photo_url",
					"allergen",
					"price",
					"description"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"DishCreateInput": {
				"$ref": "#/components/schemas/Pick_DishOutput.name-or-photo_url-or-allergen-or-price-or-description_"
			},
...
			"WithHide_Partial_Pick_DishProps.name-or-photo_url-or-allergen-or-price-or-description___": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_Pick_DishProps.name-or-photo_url-or-allergen-or-price-or-description__"
					},
					{
						"properties": {
							"hide": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
			"DishUpdateInput": {
				"$ref": "#/components/schemas/WithHide_Partial_Pick_DishProps.name-or-photo_url-or-allergen-or-price-or-description___"
			},
...
			"RestaurantSummary": {
				"properties": {
					"price_average": {
						"type": "number",
						"format": "double"
					},
					"rating_count": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"price_average",
					"rating_count",
					"rating"
				],
				"type": "object"
			},
...
			"RatingDetailOutput": {
				"$ref": "#/components/schemas/Pick_RatingRecord.title-or-comment-or-rating-or-answer-or-created_at-or-replied_at_"
			},
			"RestaurantCardOutput_RatingDetailOutput_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"nullable": true
					},
					"brand_name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"header_url": {
						"type": "string",
						"nullable": true
					},
					"restaurant_type": {
						"type": "string"
					},
					"location": {},
					"summary": {
						"$ref": "#/components/schemas/RestaurantSummary"
					},
					"detail": {
						"allOf": [
							{
								"$ref": "#/components/schemas/RatingDetailOutput"
							}
						],
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"total_bookings": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"status",
					"brand_name",
					"address",
					"city",
					"header_url",
					"restaurant_type",
					"location",
					"summary",
					"detail",
					"created_at",
					"total_bookings"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_RatingsOutput.user_id-or-updated_at-or-restaurant_id_": {
				"$ref": "#/components/schemas/Pick_RatingsOutput.Exclude_keyofRatingsOutput.user_id-or-updated_at-or-restaurant_id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UserRatingOutput": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"avatar_url": {
						"type": "string"
					}
				},
				"required": [
					"firstname"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RatingRecord": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_RatingsOutput.user_id-or-updated_at-or-restaurant_id_"
					},
					{
						"properties": {
							"replied_at": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"user_info": {
								"$ref": "#/components/schemas/UserRatingOutput"
							}
						},
						"required": [
							"replied_at",
							"user_info"
						],
						"type": "object"
					}
				]
			},
			"Record_number.number_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"RatingStatsOutput": {
				"properties": {
					"rating": {
						"type": "string"
					},
					"rating_count": {
						"type": "number",
						"format": "double"
					},
					"stats": {
						"$ref": "#/components/schemas/Record_number.number_"
					}
				},
				"required": [
					"rating",
					"rating_count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Rating": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24RatingPayload_",
				"description": "Model Rating"
			},
			"RatingsOutput": {
				"$ref": "#/components/schemas/Rating"
			},
			"Pick_RatingRecord.id-or-title-or-comment-or-rating_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"comment": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"rating",
					"title",
					"comment"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"RatingUpdateRecord": {
				"$ref": "#/components/schemas/Pick_RatingRecord.id-or-title-or-comment-or-rating_"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "resdy-back",
		"version": "1.0.0",
		"description": "Rest API with Nodejs and Prisma",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Carlos Velázquez"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/authentication/current_user": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserRecord"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Authentication service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/authentication/login": {
			"post": {
				"operationId": "LogIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCredentials"
							}
						}
					}
				}
			}
		},
		"/authentication/create_user": {
			"post": {
				"operationId": "PostUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication service"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateInput"
							}
						}
					}
				}
			}
		},
		"/authentication": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateInput"
							}
						}
					}
				}
			}
		},
		"/category/{restaurant_name}": {
			"get": {
				"operationId": "GetRestautantCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WithIsUsed_CategoryProps_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Category service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Category service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryCreateInput"
							}
						}
					}
				}
			}
		},
		"/category/{category_id}": {
			"put": {
				"operationId": "PutCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Category service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "category_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryUpdateInput"
							}
						}
					}
				}
			}
		},
		"/category": {
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Category service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"category_ids": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"category_ids"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/dishes/{restaurant_name}": {
			"get": {
				"operationId": "GetDishes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DishesByCategoryOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Dishes service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/dishes/{restaurant_name}/{category_id}": {
			"post": {
				"operationId": "PostDishes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DishOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Dishes service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "category_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DishCreateInput"
							}
						}
					}
				}
			}
		},
		"/dishes/{dish_id}": {
			"put": {
				"operationId": "PutDish",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DishOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Dishes service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "dish_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DishUpdateInput"
							}
						}
					}
				}
			}
		},
		"/dishes": {
			"delete": {
				"operationId": "DeleteDish",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Dishes service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"dish_ids": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"dish_ids"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/microservices/avatar": {
			"post": {
				"operationId": "PostAvatar",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Microservices"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"image": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"image"
								]
							}
						}
					}
				}
			}
		},
		"/microservices/getSignedUrl": {
			"post": {
				"operationId": "GetSignedUrls",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Microservices"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"key": {
										"type": "string"
									}
								},
								"required": [
									"key"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/microservices/putSignedUrls": {
			"post": {
				"operationId": "PutFilesSignedUrl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"additionalProperties": {
										"type": "string"
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Microservices"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"properties": {
										"fileName": {
											"type": "string"
										},
										"contentType": {
											"type": "string"
										},
										"key": {
											"type": "string"
										}
									},
									"required": [
										"fileName",
										"contentType",
										"key"
									],
									"type": "object"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/microservices/deleteObject": {
			"post": {
				"operationId": "DeleteObject",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Microservices"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"key": {
										"type": "string"
									}
								},
								"required": [
									"key"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/rating/myRatings": {
			"get": {
				"operationId": "GetMyRating",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RestaurantCardOutput_RatingDetailOutput_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Rating service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rating/{restaurant_name}": {
			"get": {
				"operationId": "GetRatings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RatingRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Rating service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rating/stats/{restaurant_name}": {
			"get": {
				"operationId": "GetRatingStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RatingStatsOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Rating service"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "restaurant_name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/rating/{rating_id}": {
			"put": {
				"operationId": "PutRating",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RatingsOutput"
								}
							}
						}
					}
				},
				"tags": [
					"Rating service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "rating_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RatingUpdateRecord"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}